Amend Playable typeclass definition
The existing definition has toMelody defined as:

   toMelody :: p -> Melody

which is sufficient for generating Melody from a Midi Reecording.  This is now changing to:

   toMelody :: p -> InstrumentChannels -> Melody

which is required for other sound sources which may need to generate MIDI note channel numbers by looking up the instrument names (which is the case for PSoM sources).  Sources that don't need this argument may simply pass in empty.